chore: move tests next to implementation
