+++ packages/material-ui-styles/src/makeStyles/makeStyles.spec.tsx
@@ -2,7 +2,6 @@ import * as React from 'react';
 import { Theme } from '@material-ui/core';
 import { AppBarProps } from '@material-ui/core/AppBar';
 import { createStyles, makeStyles } from '@material-ui/styles';
-import styled, { StyledProps } from '@material-ui/styles/styled';
 
 // makeStyles
 {
@@ -135,51 +134,4 @@ import styled, { StyledProps } from '@material-ui/styles/styled';
     // This doesn't fail, because inferrence is broken
     classes.other;
   }
-}
-
-// styled
-{
-  const StyledButton = styled('button')({
-    background: 'linear-gradient(45deg, #FE6B8B 30%, #FF8E53 90%)',
-    borderRadius: 3,
-    border: 0,
-    color: 'white',
-    height: 48,
-    padding: '0 30px',
-    boxShadow: '0 3px 5px 2px rgba(255, 105, 135, .3)',
-  });
-  const renderedStyledButton = <StyledButton classes={{ root: 'additional-root-class' }} />;
-  // $ExpectError
-  const nonExistingClassKey = <StyledButton classes={{ notRoot: 'additional-root-class' }} />;
-
-  interface MyTheme {
-    fontFamily: string;
-  }
-  const MyThemeInstance: MyTheme = {
-    fontFamily: 'monospace',
-  };
-  // tslint:disable-next-line: no-empty-interface
-  interface MyComponentProps extends StyledProps {
-    defaulted: string;
-  }
-  class MyComponent extends React.Component<MyComponentProps> {
-    static defaultProps = {
-      defaulted: 'Hello, World!',
-    };
-    render() {
-      const { className, defaulted } = this.props;
-      return <div className={className}>Greeted?: {defaulted.startsWith('Hello')}</div>;
-    }
-  }
-  const StyledMyComponent = styled<typeof MyComponent>(MyComponent)(
-    ({ theme }: { theme: MyTheme }) => ({
-      fontFamily: theme.fontFamily,
-    }),
-  );
-  const renderedMyComponent = (
-    <React.Fragment>
-      <MyComponent className="test" />
-      <StyledMyComponent theme={MyThemeInstance} />
-    </React.Fragment>
-  );
 }
